# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Runiq API
  description: API specification for Runiq - High Performance Go Background Job Queue Dashboard & Management Services.
  version: 2.6.0
paths:
  /api/stats:
    get:
      summary: Get queue statistics
      description: Returns real-time metrics including queues, active, failed, completed, and pending jobs.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueStats'
        '500':
          description: Internal Server Error

  /api/stats/stream:
    get:
      summary: Stream real-time queue statistics
      description: Server-Sent Events (SSE) endpoint that streams QueueStats updates periodically.
      responses:
        '200':
          description: SSE Stream
          content:
            text/event-stream:
              schema:
                type: string
                example: "data: {\"queues\":[]}"

  /api/jobs:
    get:
      summary: Get jobs list
      description: Returns a paginated list of jobs filtered by search query or status.
      parameters:
        - name: q
          in: query
          description: Search term (job name, arguments, ID, or trace ID)
          required: false
          schema:
            type: string
        - name: status
          in: query
          description: Filter jobs by status (pending, running, failed, completed, cancelled)
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of records per page
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobsResponse'
        '500':
          description: Internal Server Error

  /api/jobs/detail:
    get:
      summary: Get job details
      description: Returns detailed information for a single job by its ID.
      parameters:
        - name: id
          in: query
          description: The unique identifier of the job
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEnvelope'
        '400':
          description: Bad Request (missing ID)
        '404':
          description: Job Not Found
        '500':
          description: Internal Server Error

  /api/jobs/retry:
    post:
      summary: Retry a specific job
      parameters:
        - name: id
          in: query
          description: The unique identifier of the job to retry
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /api/jobs/cancel:
    post:
      summary: Cancel a specific running or pending job
      parameters:
        - name: id
          in: query
          description: The unique identifier of the job to cancel
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /api/jobs/failed/retry:
    post:
      summary: Retry all failed jobs
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error

  /api/jobs/failed/purge:
    post:
      summary: Purge all failed jobs
      responses:
        '200':
          description: OK
        '500':
          description: Internal Server Error

  /api/jobs/bulk-retry:
    post:
      summary: Retry multiple jobs by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkRequest'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /api/jobs/bulk-cancel:
    post:
      summary: Cancel multiple jobs by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkRequest'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /api/jobs/bulk-purge:
    post:
      summary: Purge multiple jobs by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkRequest'
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /api/queues/clear:
    post:
      summary: Clear all jobs from a queue
      parameters:
        - name: name
          in: query
          description: Name of the queue to clear
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /api/queues/pause:
    post:
      summary: Pause polling on a queue
      parameters:
        - name: name
          in: query
          description: Name of the queue to pause
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /api/queues/resume:
    post:
      summary: Resume polling on a paused queue
      parameters:
        - name: name
          in: query
          description: Name of the queue to resume
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /metrics:
    get:
      summary: Prometheus metrics endpoint
      description: Returns standard queue/jobs metrics in Prometheus exposition format.
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: "runiq_jobs_count{queue=\"default\", status=\"pending\"} 0"

components:
  schemas:
    QueueStats:
      type: object
      properties:
        queues:
          type: array
          items:
            $ref: '#/components/schemas/QueueDetail'
    QueueDetail:
      type: object
      properties:
        name:
          type: string
        pending:
          type: integer
        running:
          type: integer
        failed:
          type: integer
        processed:
          type: integer
        paused:
          type: boolean
    JobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobEnvelope'
        total:
          type: integer
    JobEnvelope:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        queue:
          type: string
        args:
          type: string
          description: JSON encoded payload arguments
        status:
          type: string
        run_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        max_retries:
          type: integer
        retries:
          type: integer
        last_error:
          type: string
        trace_id:
          type: string
    BulkRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          items:
            type: string